Add SQLite schema consumer (#40)
SQLDelight is a popular multiplatform ORM that is able to expose its schemas via a Gradle task. On #19, we discuss the possibility of ingesting them to generate ER diagrams similarly to how we do for Android's Room.

Interestingly, they export `.db` files, standard SQLite database definitions, that can be connected to and queried. Here, we rely heavily on the `PRAGMA` functions to introspect into the database schema and pull out the interesting properties to build our `DBML` memory model.

This PR, then, wires up the new `SqliteConsumer`  (with the help of a input file format sniffer heuristic) to allow not only for integration between FloorPlan and SQLDelight, but with every other ORM that can expose `.db` files (potentially even [iOS' CoreData](https://stackoverflow.com/a/15119886/12511149)).